fix error(at set content view   of side bar java) change text color "문의하기" at setting_askenroll.xml
todo: set onclick method on navigationIcon of toolbars
make swith button on alarm.xml
