+++ jibcon/src/main/java/com/sm_arts/jibcon/MainActivity.java
@@ -28,7 +28,7 @@ import com.bumptech.glide.Glide;
 import com.sm_arts.jibcon.Cheatkey.TrickMenuActivity;
 import com.sm_arts.jibcon.Conshop.MarketMenuActivity;
 import com.sm_arts.jibcon.Device.DeviceMenuActivity;
-import com.sm_arts.jibcon.settings.SettingsActivity;
+import com.sm_arts.jibcon.setting.SettingActivity;
 
 import com.sm_arts.jibcon.sidebar.AboutJibconActivity;
 import com.sm_arts.jibcon.sidebar.ConnectedDevicesActivity;
@@ -220,7 +220,7 @@ public class MainActivity extends AppCompatActivity
         mtoSettingBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent=new Intent(MainActivity.this,SettingsActivity.class);
+                Intent intent=new Intent(MainActivity.this,SettingActivity.class);
                 startActivity(intent);
             }
         });
