+++ jibcon/src/main/java/com/sm_arts/jibcon/sidebar/ConnectedDevicesActivity.java
@@ -21,7 +21,7 @@ public class ConnectedDevicesActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
 
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.sidebar_about_jibcon);
+        setContentView(R.layout.sidebar_connected_devices);
 
         /* add String[] to ListView*/
         ArrayAdapter adapter = new ArrayAdapter(this,android.R.layout.simple_list_item_1, mSidebarConnectedDeviceList);
