+++ jibcon/src/main/res/layout/setting_askenrollcompany.xml
@@ -98,6 +98,7 @@
             android:layout_height="0dp"
             android:layout_weight="1"
             android:background="#36C3FF"
+            android:textColor="@color/white"
             android:text="문의하기" />
     </LinearLayout>
