+++ jibcon/src/main/res/layout/sidebar_about_jibcon.xml
@@ -21,14 +21,20 @@
 
     <ImageView
         android:id="@+id/imageView4"
-        android:layout_width="match_parent"
+        android:layout_gravity="center"
+        android:layout_width="200dp"
         android:layout_height="wrap_content"
         app:srcCompat="@drawable/jibcon_logo_en" />
 
     <TextView
         android:id="@+id/textView4"
-        android:layout_width="match_parent"
+        android:layout_gravity="center"
+        android:layout_width="200dp"
         android:layout_height="wrap_content"
-        android:text="TextView" />
+        android:textSize="20dp"
+        android:textAlignment="center"
+        android:text=
+            "처음 만나는 스마트홈\n쉬운 스마트홈\n재밌는 스마트홈\n\n우리가 만들어갑니다"
+        android:textColor="@color/colorAccent"/>
 
 </LinearLayout>
