+++ jibcon/src/main/res/layout/sidebar_user_authority.xml
@@ -18,7 +18,7 @@
     </android.support.v7.widget.Toolbar>
 
     <ListView
-        android:id="@+id/listview_user_authority"
+        android:id="@+id/Lv_user_authority"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 </LinearLayout>
